Split up iterators some more
