+++ rust/src/iterators/rotations.rs
@@ -1,3 +1,6 @@
+//! This module implements an iterator that yeidls all of the rotations
+//! of a polycube.
+
 use crate::PolyCube;
 
 struct PlaneIterator {
@@ -104,111 +107,6 @@ impl PolyCube {
         AllRotationsIter {
             inner: rots_in_native_plane.chain(all_others),
             rotations_checked: 0,
-        }
-    }
-}
-
-type Sides = std::array::IntoIter<(usize, usize, usize), 6>;
-
-#[derive(Clone)]
-struct ToggleIterator {
-    dim_1: usize,
-    dim_2: usize,
-    dim_3: usize,
-    iterating_cube: Option<Sides>,
-    padded_cube: PolyCube,
-    done: bool,
-}
-
-impl ToggleIterator {
-    /// Move to the next possibly-occupied box
-    ///
-    /// Returns `true` if we're done iterating
-    fn go_to_next(&mut self) -> bool {
-        if self.dim_1 == self.padded_cube.dim_1 - 1 && self.dim_2 == self.padded_cube.dim_2 - 1 {
-            self.dim_1 = 1;
-            self.dim_2 = 1;
-            self.dim_3 += 1;
-        } else if self.dim_1 == self.padded_cube.dim_1 - 1 {
-            self.dim_1 = 1;
-            self.dim_2 += 1;
-        } else {
-            self.dim_1 += 1;
-        }
-
-        self.dim_3 == self.padded_cube.dim_3 - 1
-    }
-
-    fn faces(d1: usize, d2: usize, d3: usize) -> std::array::IntoIter<(usize, usize, usize), 6> {
-        [
-            (d1 + 1, d2, d3),
-            (d1 - 1, d2, d3),
-            (d1, d2 + 1, d3),
-            (d1, d2 - 1, d3),
-            (d1, d2, d3 + 1),
-            (d1, d2, d3 - 1),
-        ]
-        .into_iter()
-    }
-}
-
-impl Iterator for ToggleIterator {
-    type Item = PolyCube;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            if self.done && self.iterating_cube.is_none() {
-                return None;
-            }
-
-            if self.iterating_cube.is_none() {
-                loop {
-                    let (d1, d2, d3) = (self.dim_1, self.dim_2, self.dim_3);
-
-                    if self.go_to_next() {
-                        self.done = true;
-                        return None;
-                    }
-
-                    if self.padded_cube.is_set(d1, d2, d3) {
-                        self.iterating_cube = Some(Self::faces(d1, d2, d3));
-                        break;
-                    }
-                }
-            }
-
-            if let Some(ref mut sides) = self.iterating_cube {
-                let (d1, d2, d3) = sides.next().unwrap();
-
-                if sides.len() == 0 {
-                    self.iterating_cube.take();
-                }
-
-                // If the cube is already set, skip this face
-                if self.padded_cube.is_set(d1, d2, d3) {
-                    continue;
-                }
-
-                self.padded_cube.increase_alloc_count();
-                let mut next_cube = self.padded_cube.clone();
-                next_cube.set(d1, d2, d3).unwrap();
-                return Some(next_cube);
-            } else {
-                return None;
-            }
-        }
-    }
-}
-
-impl PolyCube {
-    pub fn expand(&self) -> impl Iterator<Item = PolyCube> + Clone {
-        ToggleIterator {
-            dim_1: 1,
-            dim_2: 1,
-            dim_3: 1,
-            iterating_cube: None,
-            padded_cube: self.pad_one(),
-            done: false,
         }
     }
 }
